requirements.txt
llama-index-program-openai
llama-index-llms-llama-api
llama_index
huggingface_hub
python-dotenv
streamlit